removing_users

+++ src/main/webapp/js/admin/users.js
@@ -1,4 +1,4 @@
-function getUsers(){
+function getUsers() {
     $list = $('.table tbody');
     $.ajax({
         url: 'http://localhost:8080/user',
@@ -12,25 +12,32 @@ function getUsers(){
                     '<td>' + res[i].firstname + '</td>' +
                     '<td>' + res[i].lastname + '</td>' +
                     '<td>' + res[i].email + '</td>' +
-                    '<td><button class="btn btn-danger btn-xs btn-delete" onclick="deleteUser(id)" id='+res[i].id+'>Delete</button></td>' +
+                    '<td><button class="btn btn-danger btn-xs btn-delete" onclick="deleteUser(id)" id=' + res[i].id + '>Delete</button></td>' +
                     '</tr>');
             })
         })
 }
-function deleteUser(id){
+
+function deleteUser(id) {
     let id2 = this.id;
     let row = $(this);
-    $.ajax({
-        url: 'http://localhost:8080/user/'+id,
-        type: 'DELETE',
-        contentType:'application/json',
-        dataType: 'text',
-        success: function(result) {
-            row.closest("tr").remove();
-            alert('user: '+id+' deleted');
-        },
-        error: function(result) {
-            alert('cannot delete user: '+id);
-        }
-    })
+    let loggedUserId = sessionStorage.getItem('loggedUserId');
+    if (loggedUserId === id)
+        alert('admin cannot delete his account');
+    else {
+        $.ajax({
+            url: 'http://localhost:8080/user/' + id,
+            type: 'DELETE',
+            contentType:'application/json',
+            dataType: 'text',
+            success: function () {
+                alert('user: ' + id + ' deleted')
+                $("#getUsers").click();
+            },
+            error: function () {
+                alert('cannot delete user: ' + id);
+            }
+        });
+    }
+
 }
